Move all API connectors to separate directory
